Flyttade market logiken till en separat komponent
/store link skafade
